Move utility functions to vk_utils module
... and improve the documents and tests.
